# Optimise topk where k==1